Added tests for the examples.
