+++ examples/validation/case2.py
@@ -17,65 +17,25 @@
     You should have received a copy of the GNU Lesser General Public License
     along with StanShock.  If not, see <https://www.gnu.org/licenses/>.
 '''
-from StanShock.stanShock import stanShock
+import time
+from typing import Optional
+import os
+
 import numpy as np
 import matplotlib as mpl
 from matplotlib import pyplot as plt
-import time
 import cantera as ct
 
-
-def isFloat(value):
-    '''
-    function isfloat
-    ==========================================================================
-    hacky python way to detect floats
-    ''' 
-    try:
-        float(value)
-        return True
-    except ValueError:
-        return False
-
-def loadData(fileName):
-    '''
-    function loadData
-    ==========================================================================
-    This function loads the raw data contained in the csv file and initiates
-    a list of dictionaries containg the data
-        fileName: file name of csv data
-        Return: list of dictionaries for each example
-    '''
-    import csv
-    rawData=[]
-    with open(fileName) as csvFile:
-        reader = csv.DictReader(csvFile)
-        for dictRow in reader:
-            cleanDictRow = {key:(float(dictRow[key]) if isFloat(dictRow[key]) else dictRow[key]) for key in dictRow}
-            rawData.append(cleanDictRow)
-    return rawData
-def getPressureData(fileName):
-    '''
-    function getPressureData
-    ==========================================================================
-    This function returns the formatted pressure vs time data
-        Inputs:
-            fileNamefile name of csv data
-        Outputs: 
-             t = time [s]
-             p = pressure [Pa]
-    '''   
-    rawData = loadData(fileName)
-    t = np.array([example["Time (s)"] for example in rawData])
-    p = np.array([example["Pressure (atm)"] for example in rawData])
-    p *= ct.one_atm
-    return (t,p)
+from StanShock.stanShock import stanShock
+from StanShock.utils import getPressureData
 
 
-if __name__ == "__main__":
+def main(data_filename: str = "data/validation/case2.csv",
+         mech_filename: str = "data/mechanisms/Nitrogen.xml",
+         show_results: bool = True,
+         results_location: Optional[str] = None) -> None:
     #=============================================================================
     #provided condtions for Case 2
-    fileName = "case2.csv"
     Ms = 2.518914
     T1 = 291.75
     p1 = 2026.499994
@@ -94,9 +54,8 @@ if __name__ == "__main__":
     #Set up gasses and determine the initial pressures
     u1 = 0.0;
     u4 = 0.0; #initially 0 velocity
-    mech="Nitrogen.xml"
-    gas1 = ct.Solution(mech)
-    gas4 = ct.Solution(mech)
+    gas1 = ct.Solution(mech_filename)
+    gas4 = ct.Solution(mech_filename)
     T4 = T1; #assumed
     gas1.TP = T1,p1
     gas4.TP = T4,p1 #use p1 as a place holder
@@ -137,13 +96,13 @@ if __name__ == "__main__":
     state1 = (gas1,u1)
     state4 = (gas4,u4)
     ssbl = stanShock(gas1,initializeRiemannProblem=(state4,state1,geometry),
-                       boundaryConditions=boundaryConditions,
-                       cfl=.9,
-                       outputEvery=100,
-                       includeBoundaryLayerTerms=True,
-                       Tw=T1, #assume wall temperature is in thermal eq. with gas
-                       DOuter= D,
-                       dlnAdx=dlnAdx)
+                     boundaryConditions=boundaryConditions,
+                     cfl=.9,
+                     outputEvery=100,
+                     includeBoundaryLayerTerms=True,
+                     Tw=T1, #assume wall temperature is in thermal eq. with gas
+                     DOuter= D,
+                     dlnAdx=dlnAdx)
     ssbl.addProbe(max(ssbl.x)) #end wall probe
 
     #Solve
@@ -158,12 +117,12 @@ if __name__ == "__main__":
     gas1.TP = T1,p1
     gas4.TP = T4,p4
     ssnbl = stanShock(gas1,initializeRiemannProblem=(state4,state1,geometry),
-                       boundaryConditions=boundaryConditions,
-                       cfl=.9,
-                       outputEvery=100,
-                       includeBoundaryLayerTerms=False,
-                       DOuter= D,
-                       dlnAdx=dlnAdx)
+                      boundaryConditions=boundaryConditions,
+                      cfl=.9,
+                      outputEvery=100,
+                      includeBoundaryLayerTerms=False,
+                      DOuter= D,
+                      dlnAdx=dlnAdx)
     ssnbl.addProbe(max(ssnbl.x)) #end wall probe
 
     #Solve
@@ -173,7 +132,7 @@ if __name__ == "__main__":
     print("The process took ", t1-t0)
 
     #import shock tube data
-    tExp, pExp = getPressureData(fileName)
+    tExp, pExp = getPressureData(data_filename)
     timeDifference = (12.211-8.10)/1000.0 #difference between the test data and simulation times
     tExp+=timeDifference
 
@@ -190,4 +149,19 @@ if __name__ == "__main__":
     plt.ylabel("$p\ [\mathrm{bar}]$")
     plt.legend(loc="lower right")
     plt.tight_layout()
-    plt.show()
+    if show_results:
+        plt.show()
+
+    if results_location is not None:
+        np.savez(
+            os.path.join(results_location, "case2.npz"),
+            pressure_with_boundary_layer=ssbl.probes[0].p,
+            pressure_without_boundary_layer=ssnbl.probes[0].p,
+            time_with_boundary_layer=ssbl.probes[0].t,
+            time_without_boundary_layer=ssnbl.probes[0].t
+        )
+        plt.savefig(os.path.join(results_location, "case2.png"))
+
+
+if __name__ == "__main__":
+    main()
