--- validationCases/case1/Nitrogen.xml
@@ -1,51 +1,0 @@
-<?xml version="1.0"?>
-<ctml>
-  <validate reactions="yes" species="yes"/>
-
-  <!-- phase gas     -->
-  <phase dim="3" id="gas">
-    <elementArray datasrc="elements.xml">N </elementArray>
-    <speciesArray datasrc="#species_data">
-      N2 
-      </speciesArray>
-    <reactionArray datasrc="#reaction_data"/>
-    <state>
-      <temperature units="K">300.0</temperature>
-      <pressure units="Pa">101325.0</pressure>
-    </state>
-    <thermo model="IdealGas"/>
-    <kinetics model="GasKinetics"/>
-    <transport model="Mix"/>
-  </phase>
-
-  <!-- species definitions     -->
-  <speciesData id="species_data">
-    <!-- species N2    -->
-    <species name="N2">
-      <atomArray>N:2 </atomArray>
-      <note>121286</note>
-      <thermo>
-        <NASA Tmax="1000.0" Tmin="300.0" P0="100000.0">
-           <floatArray name="coeffs" size="7">
-             3.298677000E+00,   1.408240000E-03,  -3.963222000E-06,   5.641515000E-09, 
-             -2.444855000E-12,  -1.020900000E+03,   3.950372000E+00</floatArray>
-        </NASA>
-        <NASA Tmax="5000.0" Tmin="1000.0" P0="100000.0">
-           <floatArray name="coeffs" size="7">
-             2.926640000E+00,   1.487977000E-03,  -5.684761000E-07,   1.009704000E-10, 
-             -6.753351000E-15,  -9.227977000E+02,   5.980528000E+00</floatArray>
-        </NASA>
-      </thermo>
-      <transport model="gas_transport">
-        <string title="geometry">linear</string>
-        <LJ_welldepth units="K">97.530</LJ_welldepth>
-        <LJ_diameter units="A">3.621</LJ_diameter>
-        <dipoleMoment units="Debye">0.000</dipoleMoment>
-        <polarizability units="A3">1.760</polarizability>
-        <rotRelax>4.000</rotRelax>
-      </transport>
-    </species>
-  </speciesData>
-  <reactionData id="reaction_data">
-  </reactionData>
-</ctml>
